% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_obs_function.R
\name{avg_obs_function}
\alias{avg_obs_function}
\title{Returns average of observed values}
\usage{
avg_obs_function(obs, sim)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{sim}{A vector containing simulated values}
}
\description{
Returns average of observed values
}
