% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_sim_function.R
\name{avg_sim_function}
\alias{avg_sim_function}
\title{Returns average of simulated values}
\usage{
avg_sim_function(obs, sim)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{sim}{A vector containing simulated values}
}
\description{
Returns average of simulated values
}
