% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_function.R
\name{bias_function}
\alias{bias_function}
\title{Returns Bias}
\usage{
bias_function(obs, pred)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{pred}{A vector containing simulated values}
}
\description{
Returns Bias
}
