% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_var_df_P_outlet.R
\name{create_var_df_P_outlet}
\alias{create_var_df_P_outlet}
\title{Creates a dataframe containing subset of variables required
to calculate Phosphorus at the outlet}
\usage{
create_var_df_P_outlet(all_var_df, ebe_var_df, Obs_P)
}
\arguments{
\item{all_var_df}{Name of the dataframe containing all the simulated variables}

\item{ebe_var_df}{Name of the event by event dataframe}

\item{Obs_P}{Name of the dataframe conining observed P data}
}
\value{
A merged with all variables required for P calculations at the outlet
}
\description{
This function reads merged all variables file and a event
by event file along with the observed phosphorus data
of same length/time period and returns a dataframe
}
