% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_function.R
\name{DV_function}
\alias{DV_function}
\title{Returns DV coefficient}
\usage{
DV_function(Qobs, Qsim)
}
\arguments{
\item{Qobs}{A vector containing observed values}

\item{Qsim}{A vector containing simulated values}
}
\value{
DV
}
\description{
Returns DV coefficient
}
