% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_P.R
\name{est_P}
\alias{est_P}
\title{Estimates Phosphorus load, phosphorus in runoff,
baseflow and lateralflow}
\usage{
est_P(totalwatsed_df, par)
}
\arguments{
\item{totalwatsed_df}{dataframe containing totalwatsed file}

\item{par}{Vector containg parameter values (mg/l) in the specific order -
P_conc_Runoff, P_conc_Lateral, P_conc_Baseflow, P_conc_Sediment}
}
\value{
A dataframe containg estimated phosporus load,
 phosphorus in runoff, baseflow and lateralflow
}
\description{
This function reads in totalwatsed file and a vector containing
paramter values, estimates Phosophorus and return
a dataframe
}
