% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_GOF_yearly_N.R
\name{eval_GOF_yearly_N}
\alias{eval_GOF_yearly_N}
\title{Returns a goodness of fit statistics for WY Nitrogen}
\usage{
eval_GOF_yearly_N(df)
}
\arguments{
\item{df}{dataframe containing observed and sumulated data for Nitrogen.}
}
\value{
Goodness of Fit between observed and simulated data
}
\description{
Takes in a dataframe containing both observed and simulated variables
 (of identical length) and returns NSE, DV, KGE, MAE, RMSE, R2 and R2adj
 for each variable namely.
}
\details{
Example variable nomenclature: Observed X --> X_Load_tonnes (i.e. for total nitrogen ---> TN_Load_tonnes)
                                Simulated X --> SimX_tonnes (i.e. for total nitrogen ---> SimTN_tonnes)
}
