% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hillslope_area.R
\name{get_hillslope_area}
\alias{get_hillslope_area}
\title{Read master pass file and returns a dataframe containing area(ha) of all the simulated
hillslopes.}
\usage{
get_hillslope_area(pass_file_path, Number_of_Hillslopes_simulated)
}
\arguments{
\item{pass_file_path}{A string pointing to the location of
the master pass file from the WEPP simulation}

\item{Number_of_Hillslopes_simulated}{Numeric value specifying the number of hillslopes simulated}
}
\value{
A dataframe containing area (ha) of each simulated hillslope
}
\description{
Read master pass file and returns a dataframe containing area(ha) of all the simulated
hillslopes.
}
