% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hillslope_area_ha.R
\name{get_hillslope_area_ha}
\alias{get_hillslope_area_ha}
\title{Read master pass file and returns a dataframe containing area(ha) of all the simulated
hillslopes.}
\usage{
get_hillslope_area_ha(pass_path)
}
\arguments{
\item{pass_path}{A string pointing to the location of
the master pass file from the WEPP simulation}
}
\value{
A dataframe containing area (ha) of each simulated hillslope
}
\description{
Read master pass file and returns a dataframe containing area(ha) of all the simulated
hillslopes.
}
