% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_N_P_from_chemicalplot.R
\name{get_N_P_from_chemicalplot}
\alias{get_N_P_from_chemicalplot}
\title{This Function reads all the chemicalplot.out files from the WEPP simulation output folder,
writes hillslope number and date to each hillslope chemplot file, extracts and calculate Nitrogen
and Phosphorus variables in kg units and exports to individual csv file}
\usage{
get_N_P_from_chemicalplot(
  output_folder,
  hillslope_area_df,
  number_of_hillslopes_simulated,
  SimStartDate,
  SimEndDate
)
}
\arguments{
\item{output_folder}{A string pointing to the location of
the WEPP simulation output folder containing chemplot.out files}

\item{hillslope_area_df}{A data frame containing hillslope areas (retrived using get_hillslope_area function)}

\item{number_of_hillslopes_simulated}{A numeric value specifying number of hillslopes simulated}

\item{SimStartDate}{A string specifying simulation start date as "YYYY-MM-DD" (eg: "1990-01-01")}

\item{SimEndDate}{A string specifying simulation end date as "YYYY-MM-DD" (eg: "2010-12-31")}
}
\value{
A csv file containing all the processed values for each hillslope written
in the output_folder
}
\description{
This Function reads all the chemicalplot.out files from the WEPP simulation output folder,
writes hillslope number and date to each hillslope chemplot file, extracts and calculate Nitrogen
and Phosphorus variables in kg units and exports to individual csv file
}
