% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WatershedArea_m2.R
\name{get_WatershedArea_m2}
\alias{get_WatershedArea_m2}
\title{Extracts watershed area from the loss file}
\usage{
get_WatershedArea_m2(file)
}
\arguments{
\item{file}{A string point to the path of the loss file}
}
\value{
A numeric vector of length 1 containing the watershed area
}
\description{
This function reads the loss file, extracts and
returns area of the simulated watersehd in square meters
}
