% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.R
\name{GOF}
\alias{GOF}
\title{Returns a goodness of fit statistics for Streamflow, SWE, Phosphorus
and sediment.}
\usage{
GOF(df)
}
\arguments{
\item{df}{dataframe containing observed and sumulated data for  Streamflow, SWE, Phosphorus and sediment.}
}
\value{
Goodness of Fit between observed and simulated data
}
\description{
Takes in a dataframe containing both observed and simulated variables
 (of identical length) and returns NSE, DV, KGE, MAE, RMSE, R2 and R2adj
 for each variable namely, Streamflow, SWE, Phosphorus and sediment.
}
