% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae_function.R
\name{mae_function}
\alias{mae_function}
\title{Returns MAE}
\usage{
mae_function(obs, pred)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{pred}{A vector containing simulated values}
}
\description{
Returns MAE
}
