% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_by_WY.R
\name{mean_by_WY}
\alias{mean_by_WY}
\title{Creates a WY mean df from daily df}
\usage{
mean_by_WY(dailydf)
}
\arguments{
\item{dailydf}{A merged daily dataframes containing all the simulated varibles}
}
\value{
A dataframe containing WY mean summary
}
\description{
This function takes in the merged daily dataframe
containing all the variables from channel,  totalwatsed and
event by event dataframes and returns WY mean df
}
