% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_by_Year.R
\name{mean_by_Year}
\alias{mean_by_Year}
\title{Creates a annual mean df from daily df}
\usage{
mean_by_Year(dailydf)
}
\arguments{
\item{dailydf}{A merged daily dataframes containing all the simulated varibles}
}
\value{
A dataframe containing annual mean summary
}
\description{
This function takes in the merged daily dataframe
containing all the variables from channel,  totalwatsed and
event by event dataframes and returns annual mean df
}
