% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPairs_function.R
\name{nPairs_function}
\alias{nPairs_function}
\title{Returns number of complete pairs}
\usage{
nPairs_function(obs, pred)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{pred}{A vector containing simulated values}
}
\description{
Returns number of complete pairs
}
