% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSeff_function.R
\name{NSeff_function}
\alias{NSeff_function}
\title{Returns Nash–Sutcliffe model efficiency coefficient}
\usage{
NSeff_function(Qobs, Qsim)
}
\arguments{
\item{Qobs}{A vector containing observed values}

\item{Qsim}{A vector containing simulated values}
}
\value{
NS
}
\description{
Returns Nash–Sutcliffe model efficiency coefficient
}
