% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSeff_log_function.R
\name{NSeff_log_function}
\alias{NSeff_log_function}
\title{Returns Nash–Sutcliffe model efficiency coefficient of logarithmic data}
\usage{
NSeff_log_function(Qobs, Qsim)
}
\arguments{
\item{Qobs}{A vector containing observed values}

\item{Qsim}{A vector containing simulated values}
}
\value{
NS
}
\description{
Returns Nash–Sutcliffe model efficiency coefficient of logarithmic data
}
