% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydrograph.R
\name{plot_hydrograph}
\alias{plot_hydrograph}
\title{Returns a hydrograph of observed and simulated data and includes
precipitation for the same time period}
\usage{
plot_hydrograph(df, units = "mm")
}
\arguments{
\item{df}{Dataframe of 4 coulmns in this specific order: 1.Time/Date 2.Simulated streamflow 3.Observed streamflow 4.Precipitation}

\item{units}{units of the data}
}
\description{
Returns a hydrograph of observed and simulated data and includes
precipitation for the same time period
}
