% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs_sim_ts.R
\name{plot_obs_sim_ts}
\alias{plot_obs_sim_ts}
\title{Returns a timeseries plot of observed and simulated data}
\usage{
plot_obs_sim_ts(x, y1, y2, plot.var.name = NULL, units = NULL)
}
\arguments{
\item{x}{A vector containing time/date}

\item{y1}{A vector containing simulated values}

\item{y2}{A vector containing observed values}

\item{plot.var.name}{String containing the name of the variable being plotted}

\item{units}{units of the data}
}
\description{
Returns a timeseries plot of observed and simulated data
}
