% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ebe.R
\name{process_ebe}
\alias{process_ebe}
\title{Reads and processes event by event file}
\usage{
process_ebe(ebe_path, SimStartDate, SimEndDate)
}
\arguments{
\item{ebe_path}{A string pointing to the location of the event by event file}

\item{SimStartDate}{Start date of the simulation}

\item{SimEndDate}{End date of the simulation}
}
\value{
A event by event dataframe along with all the
sediment and Phosphours variables in tonnes
}
\description{
This function reads the event by event file and returns
a dataframe that also includes sediment and phosphorus
variables in tonnes
}
