% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_function.R
\name{R2_function}
\alias{R2_function}
\title{Returns R2}
\usage{
R2_function(obs, pred)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{pred}{A vector containing simulated values}
}
\description{
Returns R2
}
