% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ad_function.R
\name{R2ad_function}
\alias{R2ad_function}
\title{Returns adjusted  R2}
\usage{
R2ad_function(obs, pred)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{pred}{A vector containing simulated values}
}
\description{
Returns adjusted  R2
}
