% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readObsDat.R
\name{readObsDat}
\alias{readObsDat}
\title{Reads observed data into a df}
\usage{
readObsDat(fpath)
}
\arguments{
\item{fpath}{path to the csv file containing the daily observed variables}
}
\value{
a dataframe of pbserved data
}
\description{
Takes in a observed data in this specific ordered columns and format:
 column1. Date   --format = "%m/%d/%Y"
 column2. Streamflow --Units mm
 column3. Suspended sediment concentration --Units kg/day
 column4. Total Phosphorus --Units kg/day
 column5. Soluble Reactive Phosphorus --Units kg/day
 column6. Particulate Phosphorus --Units kg/day
 column7. Suspended sediment concentration --Units tonnes/day
 column8. Snow Water Equivalent --Units mm
}
