% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rmse.R, R/rmse_function.R
\name{rmse_function}
\alias{rmse_function}
\title{Calculate Root Mean Squared Error (RMSE) between simulated and observed data}
\usage{
rmse_function(obs, pred)

rmse_function(obs, pred)
}
\arguments{
\item{obs}{A vector containing observed values}

\item{pred}{A vector containing simulated values}
}
\value{
RMSE between observed and simulated values
}
\description{
Calculate Root Mean Squared Error (RMSE) between simulated and observed data

Returns RMSE
}
