% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_TP_par_at_outlet.R
\name{rmse_TP_par_at_outlet}
\alias{rmse_TP_par_at_outlet}
\title{Returns RMSE between estimated and observed total phosphorus
at the watershed outlet}
\usage{
rmse_TP_par_at_outlet(P_outlet_df, Wshed_area_m2, par)
}
\arguments{
\item{P_outlet_df}{Dataframe created by the create_var_df_P_outlet function}

\item{Wshed_area_m2}{Area of the watershed in m2}

\item{par}{Vector containg parameter values (mg/l) in the specific order-
P_conc_Runoff, P_conc_Lateral, P_conc_Baseflow, P_conc_Sediment}
}
\value{
rmse between observed and simulated Total phosphorus at the watershed outlet
}
\description{
Returns RMSE between estimated and observed total phosphorus
at the watershed outlet
}
