% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_TP_par.R
\name{rmse_TP_par_hillslopes}
\alias{rmse_TP_par_hillslopes}
\title{Returns RMSE between estimated and observed total phosphorus
for hillslopes}
\usage{
rmse_TP_par_hillslopes(totalwatsed_df, Observed_P_kg_day, par)
}
\arguments{
\item{totalwatsed_df}{dataframe containing totalwatsed file}

\item{Observed_P_kg_day}{dataframe containing}

\item{par}{Vector containg parameter values (mg/l) in the specific order-
P_conc_Runoff, P_conc_Lateral, P_conc_Baseflow, P_conc_Sediment}
}
\value{
rmse between observed and simulated Total phosphorus
}
\description{
Returns RMSE between estimated and observed total phosphorus
for hillslopes
}
