% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_by_WY.R
\name{sum_by_WY}
\alias{sum_by_WY}
\title{Creates a WY sum df from daily df}
\usage{
sum_by_WY(dailydf)
}
\arguments{
\item{dailydf}{A merged daily dataframes containing all the simulated varibles}
}
\value{
A dataframe containing WY sum summary
}
\description{
This function takes in the merged daily dataframe
containing all the variables from channel,  totalwatsed and
event by event dataframes and returns WY sum df
}
