% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_by_Year.R
\name{sum_by_Year}
\alias{sum_by_Year}
\title{Creates a annual sum df from daily df}
\usage{
sum_by_Year(dailydf)
}
\arguments{
\item{dailydf}{A merged daily dataframes containing all the simulated varibles}
}
\value{
A dataframe containing annual sum summary
}
\description{
This function takes in the merged daily dataframe
containing all the variables from channel,  totalwatsed and
event by event dataframes and returns annual sum df
}
