% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEPPRecipes.R
\docType{package}
\name{WEPPRecipes}
\alias{WEPPRecipes}
\title{WEPPRecipes}
\description{
Contains functions to post process WEPP runs,
analyze the goodness of fit and visulize the results.
}
\author{
Chinmay Deval \email{chinmay.deval91@gmail.com}
}
